[REFACTOR] 리뷰에 맞게 수정
